Resolve R2R compilers in a separate target than resolving command line arguments for R2Rd assemblies.
This will enable a new SharedFramework SDK to reuse the built-in ReadyToRun targets and still support dotnet/runtime, which currently uses a live crossgen to crossgen their shared framework.
